build: add "macro" feature to tokio dependency
In `src/channel/state`, we use the `tokio::select!` macro,
which is only available with the "macro" feature. The build has passed
fine in-repo because the Cargo.lock contains a reference to it, and it
is pulled in. If one attempts to use appinsights-rs as a dependency to
a crate without the tokio included, with the feature, it is not
appropriately built, and building appinsights-rs fails.