feat: `RemoteDependencyTelemetry` mutable timestamp
Add functionality for specifying the timestamp of a `RemoteDependencyTelemetry`, instead of always using `Utc::now()`.

This enables setting the timestamp of the request to e.g. when the dependency call was initiated, instead of when the telemetry item was submitted (which normally would be when the dependency processing was finished).