feat: `ExceptionTelemetry` mutable timestamp
Add functionality for specifying the timestamp of a `ExceptionTelemetry`, instead of always using `Utc::now()`.

Useful as preparing all the metadata that will be sent with the exception, such as a stacktrace,
might take a considerable amount of time. The timestamp of the telemetry item will then
have a significant offset compared to the timestamp of when the event actually occurred.