feat: configurable timestamp format
Add the possibility to configure the format of event timestamps
sent to appinsights. The `timestamp_format: chrono::SecondsFormat` can
now be set on the `TelemetryConfig` struct (and passed on to the
`TelemetryContext` struct), which will then be used when generating the
string representation of all event timestamps.

TL;DR: allows setting the precision of timestamp events, which
enables sub-millisecond precision.